@@ -15,15 +15,44 @@
 # Author:
 #
 
-# Entire project uses C++11
-set(CMAKE_CXX_STANDARD 11)
+# Entire project uses C++14
+set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 set(CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
-# All options defined here
+### Dependencies
+set(OpenGL_GL_PREFERENCE GLVND)
+find_package(OpenGL)
+if(OPENGL_FOUND)
+    add_definitions(-DXR_USE_GRAPHICS_API_OPENGL)
+    message(STATUS "Enabling OpenGL support")
+elseif(BUILD_ALL_EXTENSIONS)
+    message(FATAL_ERROR "OpenGL not found")
+endif()
+
+if(NOT CMAKE_VERSION VERSION_LESS 3.7.0)
+    # Find the Vulkan headers
+    find_package(VulkanHeaders)
+    find_package(Vulkan)
+endif()
+if(VulkanHeaders_FOUND)
+    add_definitions(-DXR_USE_GRAPHICS_API_VULKAN)
+    message(STATUS "Enabling Vulkan support")
+elseif(BUILD_ALL_EXTENSIONS)
+    message(FATAL_ERROR "Vulkan headers not found")
+endif()
+
+find_package(Threads REQUIRED)
+find_package(JsonCpp)
+
+### All options defined here
 option(BUILD_LOADER "Build loader" ON)
 option(BUILD_ALL_EXTENSIONS "Build loader and layers with all extensions" OFF)
-option(BUILD_LOADER_WITH_EXCEPTION_HANDLING "Enable exception handling in the loader. Leave this on unless your standard library is built to not throw." ON)
+option(
+    BUILD_LOADER_WITH_EXCEPTION_HANDLING
+    "Enable exception handling in the loader. Leave this on unless your standard library is built to not throw."
+    ON
+)
 if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
     option(DYNAMIC_LOADER "Build the loader as a .dll library" OFF)
 else()
